Update how the full menu is generated
Use existing albums attribute on each album to get subalbums.
The albums attribute is sorted by the configured albums_sort_attr.
Add custom sorting and title to the demo gallery.
Remove need for adding the full_tree to the Gallery.
Simplify menu template to be based on albums and their subalbums.